requirements.txt
tkintermapview
customtkinter
Pillow